Replace SmartDir class with SelfDeletingFolder
