+++ test/CMakeLists.txt
@@ -38,6 +38,8 @@ target_sources(
           ${CMAKE_CURRENT_SOURCE_DIR}/units/stats/misc_test.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/units/stats/tests_test.cpp)
 
+target_include_directories(test_main PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/units/include)
+
 target_link_libraries(
   test_main
   PRIVATE project_warnings
