+++ test/test_main.cpp
@@ -6,8 +6,8 @@
 
 #include <catch2/catch.hpp>
 
-#include "modle/common/smartdir.hpp"
+#include "modle/test/self_deleting_folder.hpp"  // for SelfDeletingFolder
 
 namespace modle::test {
-[[maybe_unused]] inline const SmartDir testdir{true};  // NOLINT(cert-err58-cpp)
+[[maybe_unused]] inline const SelfDeletingFolder testdir{true};  // NOLINT(cert-err58-cpp)
 }  // namespace modle::test
